# docs/swagger.yaml
basePath: /
definitions:
...
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 多渠道 AI 聊天机器人框架的 HTTP API
  title: thinkbot API
  version: "1.0"
paths:
...
  /api/stats/bots/{id}:
    get:
      description: 查询单个 Bot 的模型用量统计
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 起始日期 (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: Bot 统计
      tags:
      - 统计
  /api/stats/bots/{id}/daily:
    get:
      description: 查询单个 Bot 的按日模型用量统计
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 起始日期 (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 按日统计
      tags:
      - 统计
  /api/stats/overview:
    get:
      description: 全平台模型用量统计概览（需要 user.manage 权限）
      parameters:
      - description: 起始日期 (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 统计概览
      tags:
      - 统计
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"